requirements.txt
flask>=3.0,<4
flask-cors>=4.0,<5
httpx>=0.27,<1
Pillow>=10.0,<12
pyzbar>=0.1.9
reportlab>=4.0,<5
gunicorn>=22.0,<24
